Consolidate header files into a single file
