--- include/impl/reverse_bits.h
@@ -1,82 +1,0 @@
-#ifndef CRC_CPP_IMPL_REVERSE_BITS_H_INCLUDED
-#define CRC_CPP_IMPL_REVERSE_BITS_H_INCLUDED
-/*
- * MIT License
- *
- * Copyright (c) 2020 Ashley Roll
- * https://github.com/AshleyRoll/crc_cpp
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-#include <cstdint>
-
-namespace crc_cpp
-{
-namespace util
-{
-    //
-    // Efficient bit reversal routines
-    //
-    template<typename T> constexpr T reverse_bits(T value);
-
-    template<> [[nodiscard]] constexpr uint8_t reverse_bits(uint8_t value)
-    {
-        value = ((value & 0xAA) >> 1) | ((value & 0x55) << 1);
-        value = ((value & 0xCC) >> 2) | ((value & 0x33) << 2);
-        value = ((value & 0xF0) >> 4) | ((value & 0x0F) << 4);
-        return value;
-    }
-
-    template<> [[nodiscard]] constexpr uint16_t reverse_bits(uint16_t value)
-    {
-        value = ((value & 0xAAAA) >> 1) | ((value & 0x5555) << 1);
-        value = ((value & 0xCCCC) >> 2) | ((value & 0x3333) << 2);
-        value = ((value & 0xF0F0) >> 4) | ((value & 0x0F0F) << 4);
-        value = (value >> 8) | (value << 8);
-        return value;
-    }
-
-    template<> [[nodiscard]] constexpr uint32_t reverse_bits(uint32_t value)
-    {
-        value = ((value & 0xAAAAAAAA) >> 1) | ((value & 0x55555555) << 1);
-        value = ((value & 0xCCCCCCCC) >> 2) | ((value & 0x33333333) << 2);
-        value = ((value & 0xF0F0F0F0) >> 4) | ((value & 0x0F0F0F0F) << 4);
-        value = ((value & 0xFF00FF00) >> 8) | ((value & 0x00FF00FF) << 8);
-        value = (value >> 16) | (value << 16);
-        return value;
-    }
-
-    template<> [[nodiscard]] constexpr uint64_t reverse_bits(uint64_t value)
-    {
-        value = ((value & 0xAAAAAAAAAAAAAAAA) >>  1) | ((value & 0x5555555555555555) <<  1);
-        value = ((value & 0xCCCCCCCCCCCCCCCC) >>  2) | ((value & 0x3333333333333333) <<  2);
-        value = ((value & 0xF0F0F0F0F0F0F0F0) >>  4) | ((value & 0x0F0F0F0F0F0F0F0F) <<  4);
-        value = ((value & 0xFF00FF00FF00FF00) >>  8) | ((value & 0x00FF00FF00FF00FF) <<  8);
-        value = ((value & 0xFFFF0000FFFF0000) >> 16) | ((value & 0x0000FFFF0000FFFF) << 16);
-        value = (value >> 32) | (value << 32);
-        return value;
-    }
-
-}   // namespace util
-}   // namespace crc_cpp
-
-
-
-#endif //CRC_CPP_IMPL_REVERSE_BITS_H_INCLUDED
